+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Changed
+
+- Removed limit on range limit for stats and log
+- Lack of end date in date range implies today (eg. 2025/08/12...)
+- "today" can be used in date range (eg. 2025/08/12...today)
+
 ## [v0.5.0] - Feb 22, 2025
 
 ### Added